% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_preprocess.R
\name{check_and_preprocess}
\alias{check_and_preprocess}
\title{Validate and preprocess the data}
\usage{
check_and_preprocess(dat, dW)
}
\arguments{
\item{dat}{The dat}

\item{dW}{The d w}
}
\description{
Validate and preprocess the data
}

