% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_positivity.R
\name{check_positivity}
\alias{check_positivity}
\title{Check whether positivity assumption is violated in data}
\usage{
check_positivity(dat, posit_level = 0.05)
}
\arguments{
\item{dat}{input data.frame}

\item{posit_level}{level of positivity to detect from data}
}
\description{
censored data NOT supported
}

