% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_update.R
\name{compute.update}
\alias{compute.update}
\title{Title}
\usage{
\method{compute}{update}(D1.t.func.prev, Pn.D1.func.prev, dat, T.uniq, W_names,
  dW)
}
\arguments{
\item{dW}{}
}
\description{
Title
}
\examples{
# TO DO
}

