% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Y.t.vec_censor.R
\name{create.Y.t.vec}
\alias{create.Y.t.vec}
\title{compute \eqn{I\{T.tilde >= t\}}}
\usage{
create.Y.t.vec(Time, t.vec)
}
\arguments{
\item{Time}{length n vector of failure time}

\item{t.vec}{t value of interest}
}
\value{
a binary vector, of length = t.vec
}
\description{
loop over t.vec
}
\examples{
# TO DO
}

