% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Y.t.vec_censor.R
\name{create.Y.t.vec_Delta1}
\alias{create.Y.t.vec_Delta1}
\title{compute I{T.tilde >= t, Delta = 1}, loop over t.vec}
\usage{
create.Y.t.vec_Delta1(Time, Delta, t.vec)
}
\arguments{
\item{Time}{length n vector of failure time}

\item{Delta}{length n vector of censoring indicator}

\item{t.vec}{t value of interest}
}
\value{
a binary vector, of length = t.vec
}
\description{
compute I{T.tilde >= t, Delta = 1}, loop over t.vec
}
\examples{
# TO DO
}

