% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep_single_t.R
\name{onestep_single_t}
\alias{onestep_single_t}
\title{Title}
\usage{
onestep_single_t(dat, tk, dW = 1, SL.trt = c("SL.glm", "SL.step",
  "SL.earth"), SL.ctime = c("SL.glm", "SL.step", "SL.earth"),
  SL.ftime = c("SL.glm", "SL.step", "SL.earth"), maxIter = 300,
  epsilon_step = 0.001, tol = 1/nrow(dat), verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{
# TO DO
}

