% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.one.step.surv.R
\name{plot.onestep.surv}
\alias{plot.onestep.surv}
\title{Plot the survival curve estimator}
\usage{
\method{plot}{onestep.surv}(onestepfit, col = "green", add = FALSE, ...)
}
\arguments{
\item{onestepfit}{object returned by surv.one.step or surv.one.step.complete}

\item{col}{line color}

\item{add}{whether to add to existing plot}

\item{lty}{line type}
}
\value{
NA
}
\description{
Plot the survival curve estimator
}

