% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survtmle_survival.R
\name{plot.survtmle_survival}
\alias{plot.survtmle_survival}
\title{Plot the survival curve estimator}
\usage{
\method{plot}{survtmle_survival}(obj, add = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
NA
}
\description{
Plot the survival curve estimator
}

