% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_onestep.R
\name{surv.one.step.complete}
\alias{surv.one.step.complete}
\title{compute one-step survival curve estimator for complete data}
\usage{
surv.one.step.complete(dat, dW, g.lib = c("SL.glm", "SL.step",
  "SL.glm.interaction"), ht.SL.Lib = c("SL.mean", "SL.glm", "SL.gam",
  "SL.earth"), ..., epsilon.step = 1e-05, max.iter = 1000,
  tol = 1/nrow(dat), verbose = TRUE)
}
\arguments{
\item{dat}{data.frame with columns T, A, W. All columns with character "W" will be treated as baseline covariates.}

\item{dW}{binary input vector specifying dynamic treatment (as a function output of W)}

\item{g.lib}{SuperLearner library for fitting treatment regression}

\item{ht.SL.Lib}{SuperLearner library for fitting conditional hazard regression}

\item{...}{additional options for plotting initial fit curve}

\item{epsilon.step}{step size for one-step recursion}

\item{max.iter}{maximal number of recursion for one-step}

\item{tol}{tolerance for optimization}

\item{verbose}{to plot the initial fit curve and the objective function value during optimzation}
}
\value{
Psi.hat vector of survival curve under intervention

T.uniq vector of time points where Psi.hat gets values (have same length as Psi.hat)

params list of meta-information of estimation

variables list of data summary

initial_fit list of initial fit (hazard, g_1, Delta)
}
\description{
options to ADD:
SL.formula: the covariates to include in SL
}

