% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle_survival.R
\name{survtmle_survival}
\alias{survtmle_survival}
\title{Iterative TMLE for survival curve}
\usage{
survtmle_survival(dat)
}
\arguments{
\item{dat}{data.frame with columns T, A, W. All columns with character "W" will be treated as baseline covariates.}
}
\value{
data.frame, where the first column is survival probability, second column is the time point of the survival curve
}
\description{
Iterative TMLE for survival curve
}

