% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_check_positivity.R
\name{check_positivity}
\alias{check_positivity}
\title{Check positivity assumption of input data}
\usage{
check_positivity(dat, posit_level = 0.05)
}
\arguments{
\item{dat}{input data.frame}

\item{posit_level}{level of positivity to detect from data}
}
\value{
raise warning when positivity violated
}
\description{
Check whether positivity assumption is violated in data
censored data NOT supported
}

