% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.step.cdf_vec.R
\name{compute.step.cdf}
\alias{compute.step.cdf}
\title{compute cdf of step function}
\usage{
\method{compute}{step.cdf}(pdf.mat, t.vec, start = -Inf)
}
\arguments{
\item{pdf.mat}{if input vector = compute cdf for a step-function pdf;
if input matrix = compute cdf for several step-function pdf with same jump points}

\item{t.vec}{unique jump points of step function}

\item{start}{if -Inf = from left to right; if Inf = from right to left.}
}
\value{
vector
}
\description{
compute cdf of step function
}
\examples{
# TO DO
}

