% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Y.t.vec_censor.R
\name{create.Y.t.vec}
\alias{create.Y.t.vec}
\title{compute I{T.tilde >= t}, loop over t.vec}
\usage{
create.Y.t.vec(Time, t.vec)
}
\arguments{
\item{t.vec}{}
}
\value{
a binary vector, of length = t.vec
}
\description{
compute I{T.tilde >= t}, loop over t.vec
}
\examples{
# TO DO
}

