% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.one.step.surv.R
\name{plot.onestep.surv}
\alias{plot.onestep.surv}
\title{Title}
\usage{
\method{plot}{onestep.surv}(onestepfit, col = "green", lty = 1,
  add = FALSE)
}
\arguments{
\item{add}{}
}
\description{
Title
}

