% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_survtmle_survival.R
\name{survtmle_survival}
\alias{survtmle_survival}
\title{Iterative TMLE for survival curve}
\usage{
survtmle_survival(dat, dW = rep(1, nrow(dat)), T.cutoff = NULL,
  SL.ftime = c("SL.glm", "SL.mean", "SL.step", "SL.earth"),
  SL.ctime = c("SL.glm", "SL.mean"), SL.trt = c("SL.glm", "SL.mean",
  "SL.step", "SL.earth"))
}
\arguments{
\item{dat}{data.frame with columns T, A, W. All columns with character "W" will be treated as baseline covariates.}
}
\value{
data.frame, where the first column is survival probability, second column is the time point of the survival curve
}
\description{
Iterative TMLE for survival curve
}

