% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle_survival_single_t.R
\name{survtmle_survival_single_t}
\alias{survtmle_survival_single_t}
\title{Iterative TMLE for survival curve at specific time point}
\usage{
survtmle_survival_single_t(dat, tk)
}
\arguments{
\item{dat}{data.frame with columns T, A, C, W. All columns with character "W" will be treated as baseline covariates.}

\item{tk}{time point to compute survival probability}
}
\description{
Iterative TMLE for survival curve at specific time point
}

